Added the product information to be persisted to the db when the application starts, updated the Product repository and model to reflect and int for the id.

+++ src/main/java/com/rachelcampbell/RedStoreSA/Model/Product.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // primary key
+    private int id; // primary key
     private String name; // name of item (will not be unique)
     private String dept; // (clothing, shoes, or accessories)
     private String type; // (tshirt, polo, athletic pants, shoes, watch, fitness tracker, etc)
